+++ tests/integration/conversations/v1/test_conversation.py
@@ -41,6 +41,10 @@ class ConversationTestCase(IntegrationTestCase):
                 "date_created": "2015-12-16T22:18:37Z",
                 "date_updated": "2015-12-16T22:18:38Z",
                 "state": "inactive",
+                "timers": {
+                    "date_inactive": "2015-12-16T22:19:38Z",
+                    "date_closed": "2015-12-16T22:28:38Z"
+                },
                 "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -82,6 +86,10 @@ class ConversationTestCase(IntegrationTestCase):
                 "date_created": "2015-12-16T22:18:37Z",
                 "date_updated": "2015-12-16T22:18:38Z",
                 "state": "inactive",
+                "timers": {
+                    "date_inactive": "2015-12-16T22:19:38Z",
+                    "date_closed": "2015-12-16T22:28:38Z"
+                },
                 "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -144,6 +152,10 @@ class ConversationTestCase(IntegrationTestCase):
                 "date_created": "2015-12-16T22:18:37Z",
                 "date_updated": "2015-12-16T22:18:38Z",
                 "state": "active",
+                "timers": {
+                    "date_inactive": "2015-12-16T22:19:38Z",
+                    "date_closed": "2015-12-16T22:28:38Z"
+                },
                 "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
                     "participants": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",
@@ -185,6 +197,10 @@ class ConversationTestCase(IntegrationTestCase):
                         "date_created": "2015-12-16T22:18:37Z",
                         "date_updated": "2015-12-16T22:18:38Z",
                         "state": "active",
+                        "timers": {
+                            "date_inactive": "2015-12-16T22:19:38Z",
+                            "date_closed": "2015-12-16T22:28:38Z"
+                        },
                         "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "links": {
                             "participants": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants",