+++ twilio/rest/numbers/v2/regulatory_compliance/supporting_document.py
@@ -268,6 +268,7 @@ class SupportingDocumentInstance(InstanceResource):
         REJECTED = "rejected"
         APPROVED = "approved"
         EXPIRED = "expired"
+        PROVISIONALLY_APPROVED = "provisionally-approved"
 
     def __init__(self, version, payload, sid=None):
         """